Updated player.html
In this example, I added an h1 heading and two paragraphs of text to provide some context for the Scratch 3.0 GUI player example. I also added a bulleted list of additional resources that might be helpful for someone who is learning about Scratch. These extra features should now be visible when you open the HTML file in a web browser.